+++ scripts/load.lua
@@ -1,5 +1,7 @@
 function loadScene(host, sceneName)
-    require('scenes.' .. sceneName .. '.behaviors')
+    eventModule = require("scripts.event")
+    resumeEvent, beginEvent, simpleMessages, beginAutoMove, doAutoMove, populateDefintions = table.unpack(eventModule)
+    populateDefintions(require('scenes.' .. sceneName .. '.behaviors'))
 
     local mapTable = require('scenes.' .. sceneName .. '.map')
     _loadScene(mapTable["backgroundPath"], mapTable["allThings"], host)
